refactor(src): split worker/supervisor.rs by responsibility
